add scss, reset css, delete bootstrap
