+++ src/main.js
@@ -4,6 +4,7 @@ import { createPinia } from 'pinia';
 import App from './App.vue';
 import router from './router';
 import 'bootstrap-icons/font/bootstrap-icons.css';
+import './styles/main.scss';
 const app = createApp(App);
 
 app.use(createPinia());
