+++ src/styles/main.scss
@@ -1,0 +1,20 @@
+@import './reset.scss';
+@import './variables.scss';
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100;300;400;500;700&display=swap');
+
+html {
+  font-size: 18px;
+  height: 100%;
+}
+
+body {
+  height: 100%;
+  background-color: $bgColor;
+  color: $mainColor;
+  font-family: 'Noto Sans TC';
+}
+
+.container {
+  width: 80%;
+  margin: 0 auto;
+}
