+++ src/styles/variables.scss
@@ -1,0 +1,10 @@
+$white :#ffffff;
+$black : #000000;
+$blue  : #0099FF;
+$orange: #FF6600;
+
+$bgColor     :$white;
+$mainColor   :#1C1C1C;
+$formBgColor :#F5F8FA;
+$secondaryTextColor: #657786;
+$borderColor :#E6ECF0;
